Trait web_server_da_orchestrator::infra_da::RunDA

pub trait RunDA<TYPES: NodeType, MEMBERSHIP: Membership<TYPES> + Debug, DANETWORK: CommunicationChannel<TYPES, Message<TYPES, NODE>, MEMBERSHIP> + Debug, QUORUMNETWORK: CommunicationChannel<TYPES, Message<TYPES, NODE>, MEMBERSHIP> + Debug, VIEWSYNCNETWORK: CommunicationChannel<TYPES, Message<TYPES, NODE>, MEMBERSHIP> + Debug, NODE: NodeImplementation<TYPES, Leaf = SequencingLeaf<TYPES>, Exchanges = SequencingExchanges<TYPES, Message<TYPES, NODE>, QuorumExchange<TYPES, SequencingLeaf<TYPES>, QuorumProposal<TYPES, SequencingLeaf<TYPES>>, MEMBERSHIP, QUORUMNETWORK, Message<TYPES, NODE>>, CommitteeExchange<TYPES, MEMBERSHIP, DANETWORK, Message<TYPES, NODE>>, ViewSyncExchange<TYPES, ViewSyncCertificate<TYPES>, MEMBERSHIP, VIEWSYNCNETWORK, Message<TYPES, NODE>>>, Storage = MemoryStorage<TYPES, SequencingLeaf<TYPES>>, ConsensusMessage = SequencingMessage<TYPES, NODE>>>where
    <TYPES as NodeType>::StateType: TestableState,
    <TYPES as NodeType>::BlockType: TestableBlock,
    SequencingLeaf<TYPES>: TestableLeaf,
    Self: Sync,
    SystemContext<TYPES, NODE>: HotShotType<TYPES, NODE>,{
    // Required methods
    fn initialize_networking<'async_trait>(
        config: NetworkConfig<TYPES::SignatureKey, <TYPES::SignatureKey as SignatureKey>::StakeTableEntry, TYPES::ElectionConfigType>
    ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn get_da_network(&self) -> DANETWORK;
    fn get_quorum_network(&self) -> QUORUMNETWORK;
    fn get_view_sync_network(&self) -> VIEWSYNCNETWORK;
    fn get_config(
        &self
    ) -> NetworkConfig<TYPES::SignatureKey, <TYPES::SignatureKey as SignatureKey>::StakeTableEntry, TYPES::ElectionConfigType>;

    // Provided methods
    fn initialize_state_and_hotshot<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, NODE>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn run_hotshot<'life0, 'async_trait>(
        &'life0 self,
        context: SystemContextHandle<TYPES, NODE>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Defines the behavior of a “run” of the network with a given configuration

Required Methods

fn initialize_networking<'async_trait>( config: NetworkConfig<TYPES::SignatureKey, <TYPES::SignatureKey as SignatureKey>::StakeTableEntry, TYPES::ElectionConfigType> ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>where Self: 'async_trait,

Initializes networking, returns self

fn get_da_network(&self) -> DANETWORK

Returns the da network for this run

fn get_quorum_network(&self) -> QUORUMNETWORK

Returns the quorum network for this run

fn get_view_sync_network(&self) -> VIEWSYNCNETWORK

Returns view sync network for this run

fn get_config( &self ) -> NetworkConfig<TYPES::SignatureKey, <TYPES::SignatureKey as SignatureKey>::StakeTableEntry, TYPES::ElectionConfigType>

Returns the config for this run

Provided Methods

fn initialize_state_and_hotshot<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = SystemContextHandle<TYPES, NODE>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Initializes the genesis state and HotShot instance; does not start HotShot consensus

Panics if it cannot generate a genesis block, fails to initialize HotShot, or cannot

get the anchored view Note: sequencing leaf does not have state, so does not return state

fn run_hotshot<'life0, 'async_trait>( &'life0 self, context: SystemContextHandle<TYPES, NODE> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Starts HotShot consensus, returns when consensus has finished